+++ src/lib.rs
@@ -1,5 +1,4 @@
 mod arguments;
-mod callback;
 mod chain;
 mod database;
 mod definitions;