+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.5] - 2024-10-29
+
+### 🚀 Features
+
+- Callback in case callback url provided
+
+### 🐛 Bug Fixes
+
+- fix error handling as a result of dep uupgrade
+- fix order withdraw transaction
+- mark order withdrawn on successful withdraw
+
 ## [0.2.4] - 2024-10-21
 
 ### ⚡ Performance